requirements.txt
streamlit>=1.31.0
faiss-cpu
sentence-transformers
PyPDF2
numpy
pandas
scikit-learn
groq
python-dotenv
